+++ src/Plugin.php
@@ -1,10 +1,10 @@
 <?php
-namespace Sms77\Craft;
+namespace Seven\Craft;
 
 use Craft;
 use craft\events\RegisterCpNavItemsEvent;
 use craft\web\twig\variables\Cp;
-use Sms77\Craft\models\Settings;
+use Seven\Craft\models\Settings;
 use yii\base\Event;
 
 class Plugin extends \craft\base\Plugin {
@@ -18,13 +18,13 @@ class Plugin extends \craft\base\Plugin {
     }
 
     private function registerTranslations() {
-        Craft::$app->i18n->translations['sms77*'] = [
+        Craft::$app->i18n->translations['seven*'] = [
             'allowOverrides' => true,
             'basePath' => $this->getBasePath() . '/translations',
             'class' => PhpMessageSource::class,
             'fileMap' => [
-                'sms77' => 'site',
-                'sms77-app' => 'app',
+                'seven' => 'site',
+                'seven-app' => 'app',
             ],
             'sourceLanguage' => 'en',
         ];
@@ -34,13 +34,13 @@ class Plugin extends \craft\base\Plugin {
         Event::on(Cp::class, Cp::EVENT_REGISTER_CP_NAV_ITEMS,
             function (RegisterCpNavItemsEvent $event) {
                 $event->navItems[] = [
-                    'label' => 'sms77 ' . Craft::t('sms77', 'SMS'),
-                    'url' => 'sms77/sms',
+                    'label' => 'seven ' . Craft::t('seven', 'SMS'),
+                    'url' => 'seven/sms',
                 ];
 
                 $event->navItems[] = [
-                    'label' => 'sms77 ' . Craft::t('sms77', 'Voice'),
-                    'url' => 'sms77/voice',
+                    'label' => 'seven ' . Craft::t('seven', 'Voice'),
+                    'url' => 'seven/voice',
                 ];
             }
         );
@@ -52,6 +52,6 @@ class Plugin extends \craft\base\Plugin {
 
     protected function settingsHtml() {
         return Craft::$app->getView()
-            ->renderTemplate('sms77/settings', ['settings' => $this->getSettings()]);
+            ->renderTemplate('seven/settings', ['settings' => $this->getSettings()]);
     }
 }